Chore: Refactored Admin Article Editor UI
