+++ src/app/components/AdminHeader.tsx
@@ -30,7 +30,7 @@ export default function AdminHeader({
 
   return (
     <>
-      <div className="fixed top-0 left-0 right-0 z-40 bg-white border-b border-gray-200">
+      <div className="fixed top-0 left-0 right-0 z-100 bg-white border-b border-gray-200">
         <div className="px-4 py-2 flex items-center justify-between">
           <div className="flex items-center gap-2">
             <img
